chore: move monorepo jest helper to the workspace package
