+++ packages/@o3r/dev-tools/src/utils/is-package-installed.ts
@@ -1,8 +1,7 @@
 /**
  * Check if an npm package is installed
- *
  * @param packageName The package to check
- * @deprecated will be removed in Otter v12.
+ * @deprecated You can use the one expose in `@o3r/schematics`, will be removed in Otter v12.
  */
 export function isPackageInstalled(packageName: string) {
   try {
