+++ packages/@o3r/workspace/src/public_api.ts
@@ -1,1 +1,1 @@
-export type {};
+export * from './utils/index';
