+++ packages/@o3r/workspace/src/utils/index.ts
@@ -1,0 +1,1 @@
+export * from './jest.helpers';
